Extract a common JMS module before implementing Jakarta JMS instrumentation
