--- instrumentation/jms-1.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jms/JmsIgnoredTypesConfigurer.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.jms;
-
-import com.google.auto.service.AutoService;
-import io.opentelemetry.javaagent.extension.ignore.IgnoredTypesBuilder;
-import io.opentelemetry.javaagent.extension.ignore.IgnoredTypesConfigurer;
-import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
-
-@AutoService(IgnoredTypesConfigurer.class)
-public class JmsIgnoredTypesConfigurer implements IgnoredTypesConfigurer {
-
-  @Override
-  public void configure(IgnoredTypesBuilder builder, ConfigProperties config) {
-    // Avoid instrumenting internal OrderedExecutor worker class
-    builder.ignoreTaskClass("org.hornetq.utils.OrderedExecutorFactory$OrderedExecutor$");
-  }
-}
