+++ instrumentation/jms/jms-1.1/javaagent-unit-tests/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
