+++ instrumentation/jms/jms-common/javaagent/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+  id("otel.javaagent-instrumentation")
+}
+
+dependencies {
+  compileOnly("com.google.auto.value:auto-value-annotations")
+  annotationProcessor("com.google.auto.value:auto-value")
+}
