+++ settings.gradle.kts
@@ -290,8 +290,9 @@ hideFromDependabot(":instrumentation:jetty:jetty-common:javaagent")
 hideFromDependabot(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:javaagent")
 hideFromDependabot(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:library")
 hideFromDependabot(":instrumentation:jetty-httpclient:jetty-httpclient-9.2:testing")
-hideFromDependabot(":instrumentation:jms-1.1:javaagent")
-hideFromDependabot(":instrumentation:jms-1.1:javaagent-unit-tests")
+hideFromDependabot(":instrumentation:jms:jms-1.1:javaagent")
+hideFromDependabot(":instrumentation:jms:jms-1.1:javaagent-unit-tests")
+hideFromDependabot(":instrumentation:jms:jms-common:javaagent")
 hideFromDependabot(":instrumentation:jmx-metrics:javaagent")
 hideFromDependabot(":instrumentation:jmx-metrics:library")
 hideFromDependabot(":instrumentation:jsf:jsf-common:javaagent")
